Speedo_meter is working with xGetSpeedValue( )
1. ABS image array size is taken 24*20 size. So the problem is solved.
2. Static image (HMI_Layout) has been deleted.
3. Paths are included in model.cpp
4. IndicationStatus_t type Global variable created in Speedometer_App.h
5. IndicationStatus_t type Global variable created in Indicator_App.h
6. In Speedometer_App.h added function declaration as mentioned below. Then only function works in model.cpp file.
#ifdef __cplusplus
extern "C" {
#endif

uint32_t xGetSpeedValue(speedDisplayMetrics_t *speedDisplayUnits, IndicationStatus_t *speedStatus);

#ifdef __cplusplus
}
#endif
7.Wild Card buffer size for Speed has been changed from 6 to 4. and kmph_mph changed form 4 to 5.

+++ eStar_Cluster_Demo/TouchGFX/gui/include/gui/model/Model.hpp
@@ -24,7 +24,7 @@ public:
     //void FuelData();
 protected:
     ModelListener* modelListener;
-    uint16_t counter;
+    uint32_t counter;
     //uint16_t FuelCounter;
     uint16_t TickCount;
     uint16_t speedcounter;
